+++ dax/internal/client/single.go
@@ -157,7 +157,7 @@ func (client *SingleDaxClient) startHealthChecks(ctx context.Context, cc *cluste
 		defer cfn()
 		var err error
 		opts := RequestOptions{}
-		opts.RetryMaxAttempts = 2
+		opts.RetryMaxAttempts = 3
 		_, err = client.endpoints(ctx, opts)
 		if err != nil {
 			cc.debugLog("Health checks failed with error " + err.Error() + " for host :: " + host.host)
@@ -464,31 +464,41 @@ func (client *SingleDaxClient) executeWithContext(ctx context.Context, op string
 		return err
 	}
 	if err = client.pool.setDeadline(ctx, t); err != nil {
-		client.pool.discard(t)
+		// If the error is just due to context cancelled or timeout
+		// then the tube is still usable because we have not written anything to tube
+		if err == ctx.Err() {
+			client.pool.put(t)
+			return err
+		}
+		// If we get error while setting deadline of tube
+		// probably something is wrong with the tube
+		client.pool.closeTube(t)
 		return err
 	}
 
 	if err = client.auth(ctx, t); err != nil {
-		client.pool.discard(t)
+		// Auth method writes in the tube and
+		// it is not guaranteed that it will be drained completely on error
+		client.pool.closeTube(t)
 		return err
 	}
 
 	writer := t.CborWriter()
 	if err = encoder(writer); err != nil {
-		// Validation errors will cause pool to be discarded as there is no guarantee
+		// Validation errors will cause connection to be closed as there is no guarantee
 		// that the validation was performed before any data was written into tube
-		client.pool.discard(t)
+		client.pool.closeTube(t)
 		return err
 	}
 	if err := writer.Flush(); err != nil {
-		client.pool.discard(t)
+		client.pool.closeTube(t)
 		return err
 	}
 
 	reader := t.CborReader()
 	ex, err := decodeError(reader)
-	if err != nil { // decode or network error
-		client.pool.discard(t)
+	if err != nil { // decode or network error - doesn't guarantee completely drained tube
+		client.pool.closeTube(t)
 		return err
 	}
 	if ex != nil { // user or server error
@@ -498,7 +508,8 @@ func (client *SingleDaxClient) executeWithContext(ctx context.Context, op string
 
 	err = decoder(reader)
 	if err != nil {
-		client.pool.discard(t)
+		// we are not able to completely drain tube
+		client.pool.closeTube(t)
 	} else {
 		client.pool.put(t)
 	}
@@ -533,7 +544,7 @@ func (client *SingleDaxClient) recycleTube(t tube, err error) {
 	if recycle {
 		client.pool.put(t)
 	} else {
-		client.pool.discard(t)
+		client.pool.closeTube(t)
 	}
 }
 func (client *SingleDaxClient) auth(ctx context.Context, t tube) error {